+++ prisma/migrations/20260228090520_tag_types/migration.sql
@@ -1,0 +1,5 @@
+-- CreateEnum
+CREATE TYPE "TagType" AS ENUM ('GENERAL', 'DIALECT', 'PROP');
+
+-- AlterTable
+ALTER TABLE "Tag" ADD COLUMN     "type" "TagType" NOT NULL DEFAULT 'GENERAL';